Back to only capturing keystrokes when an accessible editable text control has focus.
This avoids many problems with the previous mechanism that tried to monitor keystrokes globally. There's now better performance, and Wayland should be supported too.
This does now mean non-accessible applications such as Chrome are no longer supported, but I have other plans for how to deal with apps like that.